LPS-167592 NodeValidator: inline
