+++ modules/apps/portal-workflow/portal-workflow-kaleo-definition-impl/src/main/java/com/liferay/portal/workflow/kaleo/definition/internal/parser/DefaultWorkflowValidator.java
@@ -14,9 +14,13 @@
 
 package com.liferay.portal.workflow.kaleo.definition.internal.parser;
 
+import com.liferay.osgi.service.tracker.collections.map.ServiceReferenceMapperFactory;
+import com.liferay.osgi.service.tracker.collections.map.ServiceTrackerMap;
+import com.liferay.osgi.service.tracker.collections.map.ServiceTrackerMapFactory;
 import com.liferay.portal.kernel.workflow.WorkflowException;
 import com.liferay.portal.workflow.kaleo.definition.Definition;
 import com.liferay.portal.workflow.kaleo.definition.Node;
+import com.liferay.portal.workflow.kaleo.definition.NodeType;
 import com.liferay.portal.workflow.kaleo.definition.State;
 import com.liferay.portal.workflow.kaleo.definition.exception.KaleoDefinitionValidationException;
 import com.liferay.portal.workflow.kaleo.definition.parser.NodeValidator;
@@ -25,8 +29,10 @@ import com.liferay.portal.workflow.kaleo.definition.parser.WorkflowValidator;
 import java.util.Collection;
 import java.util.List;
 
+import org.osgi.framework.BundleContext;
+import org.osgi.service.component.annotations.Activate;
 import org.osgi.service.component.annotations.Component;
-import org.osgi.service.component.annotations.Reference;
+import org.osgi.service.component.annotations.Deactivate;
 
 /**
  * @author Michael C. Han
@@ -59,14 +65,29 @@ public class DefaultWorkflowValidator implements WorkflowValidator {
 		Collection<Node> nodes = definition.getNodes();
 
 		for (Node node : nodes) {
-			NodeValidator<Node> nodeValidator =
-				_nodeValidatorRegistry.getNodeValidator(node.getNodeType());
+			NodeValidator<Node> nodeValidator = _serviceTrackerMap.getService(
+				node.getNodeType());
 
 			nodeValidator.validate(definition, node);
 		}
 	}
 
-	@Reference
-	private NodeValidatorRegistry _nodeValidatorRegistry;
+	@Activate
+	protected void activate(BundleContext bundleContext) {
+		_serviceTrackerMap = ServiceTrackerMapFactory.openSingleValueMap(
+			bundleContext,
+			(Class<NodeValidator<Node>>)(Class<?>)NodeValidator.class, null,
+			ServiceReferenceMapperFactory.create(
+				bundleContext,
+				(nodeValidator, emitter) -> emitter.emit(
+					nodeValidator.getNodeType())));
+	}
+
+	@Deactivate
+	protected void deactivate() {
+		_serviceTrackerMap.close();
+	}
+
+	private ServiceTrackerMap<NodeType, NodeValidator<Node>> _serviceTrackerMap;
 
 }
